Add navigation npm package and create home, placesCard, and place components
